added calls and logic to main game loop
added calls to methods and stubbed out methods within each object for
proper game flow.  Added comments and created a setfocus method to set
the focus to the next question and scroll the page down.  Moved the
showstar and showcheck methods to the output class..  Added fields to
base classes for  score keeping.